Subida del ejercicio 12

+++ TP1BazanFabricio/src/ar/edu/unju/fi/ejercicio1/main.java
@@ -4,9 +4,9 @@ public class main {
 
 	public static void main(String[] args) {
 		int num1=8, num2=5, num3=1, num4=20, num5=9;
-		double promedio;
+		Float promedio;
 		System.out.println(num1+", "+num2+", "+num3+", "+num4+", "+num5);
-		promedio=((num1+num2+num3+num4+num5)/5);
+		promedio=(float)((num1+num2+num3+num4+num5)/5);
 		System.out.println("Su promedio es: "+promedio);
 	}
